+++ packages/internal/did-resolver/src/did-cache-memory.ts
@@ -2,7 +2,7 @@ import { Did, DidDocument } from '@gander-social-atproto/did'
 import {
   SimpleStoreMemory,
   SimpleStoreMemoryOptions,
-} from '@atproto-labs/simple-store-memory'
+} from '@gander-atproto-nest/simple-store-memory'
 import { DidCache } from './did-cache.js'
 
 const DEFAULT_TTL = 3600 * 1000 // 1 hour