+++ packages/oauth/oauth-client-browser-example/rollup.config.js
@@ -8,7 +8,7 @@ const { default: swc } = require('@rollup/plugin-swc')
 const { defineConfig } = require('rollup')
 const {
   bundleManifest,
-} = require('@atproto-labs/rollup-plugin-bundle-manifest')
+} = require('@gander-atproto-nest/rollup-plugin-bundle-manifest')
 const postcss = ((m) => m.default || m)(require('rollup-plugin-postcss'))
 const serve = ((m) => m.default || m)(require('rollup-plugin-serve'))