+++ packages/oauth/oauth-client/src/fetch-dpop.ts
@@ -1,7 +1,12 @@
 import { Key } from '@gander-social-atproto/jwk'
 import { base64url } from 'multiformats/bases/base64'
-import { Fetch, FetchContext, cancelBody, peekJson } from '@atproto-labs/fetch'
-import { SimpleStore } from '@atproto-labs/simple-store'
+import {
+  Fetch,
+  FetchContext,
+  cancelBody,
+  peekJson,
+} from '@gander-atproto-nest/fetch'
+import { SimpleStore } from '@gander-atproto-nest/simple-store'
 
 // "undefined" in non https environments or environments without crypto
 const subtle = globalThis.crypto?.subtle as SubtleCrypto | undefined